Move restore test in separate files
