+++ stable/restore/files/nuorestore
@@ -54,7 +54,7 @@ function wrapLogfile() {
   fi
 }
 
-# function - NuoDB 4.2.2+ supports new way of requesting database in-place
+# function - NuoDB 4.2+ supports new way of requesting database in-place
 # restore build into nuodocker.
 #
 function isRestoreRequestSupported() {
